+++ lib/dradis/plugins/gem_version.rb
@@ -7,11 +7,11 @@ module Dradis
 
     module VERSION
       MAJOR = 3
-      MINOR = 6
-      TINY  = 2
+      MINOR = 7
+      TINY  = 0
       PRE   = nil
 
-      STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
+      STRING = [MAJOR, MINOR, TINY, PRE].compact.join('.')
     end
   end
 end